# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dispatch controller API
  version: 0.1.0
  description: Project-scoped application deployment and operations API.
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /deployment-search:
    get:
      operationId: searchDeploymentHistory
      description: Search all retained history within accessible applications; filters apply before cursor pagination.
        Returns at most 50 records without snapshots or credentials.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  next:
                    type: string
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: project
        in: query
        schema:
          type: string
      - name: application
        in: query
        schema:
          type: string
      - name: environment
        in: query
        schema:
          type: string
      - name: target
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: revision
        in: query
        schema:
          type: string
      - name: before
        in: query
        schema:
          type: string
...
  /apps/{id}/sync:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getApplicationSync
      description: Saved runtime observation, repository sync, deployment revision, and recent reapply history. Requires project.view.
      responses:
        "200":
          description: Separate sync, drift, and health indicators. Does not contact the target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationSync" }
        "403": { $ref: "#/components/responses/Problem" }
  /apps/{id}/drift/check:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      operationId: checkApplicationDrift
      description: Manual controller-side Kubernetes/OpenShift check with a 30-second deadline. Requires project.configure.
      responses:
        "200":
          description: Saved observation. Target failures produce Unknown and preserve the last successful check time.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationSync" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /apps/{id}/reapply:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      operationId: reapplyApplication
      description: Explicitly restore the saved successful deployment after dry-run validation. Requires deployment.run. Does not adopt new Git revisions or credentials. Multi-resource writes are not transactional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deploymentId]
              properties:
                deploymentId: { type: string, minLength: 1 }
      responses:
        "200":
          description: Reapply recorded and runtime checked again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationSync" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /workflow/revisions:
    get:
      operationId: listWorkflowRevisions
      parameters:
        - { in: query, name: resourceId, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/JSONList" }
...
  /deployments/{deploymentId}:
    get:
      operationId: getDeployment
      parameters:
        - $ref: "#/components/parameters/DeploymentID"
      responses:
        "200":
          description: Deployment evidence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    DeploymentID:
      in: path
      name: deploymentId
      required: true
      schema: { type: string }
...
  responses:
    JSONList:
      description: JSON array.
      content:
        application/json:
          schema: { type: array, items: {} }
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ApplicationSync:
      type: object
      required: [appId, supported, reapplyAvailable, configuration, revision, drift, actions]
      properties:
        appId: { type: string }
        deploymentId: { type: string }
        supported: { type: boolean }
        reapplyAvailable: { type: boolean }
        configuration:
          type: object
          properties:
            state: { type: string }
            message: { type: string }
            sourceId: { type: string }
            lastSyncedAt: { type: string, format: date-time }
        revision:
          type: object
          properties:
            state: { type: string }
            applied: { type: string }
            observed: { type: string }
        drift:
          type: object
          properties:
            deploymentId: { type: string }
            state: { type: string, enum: [synced, out_of_sync, unknown] }
            health: { type: string, enum: [healthy, progressing, degraded, unknown, not_applicable] }
            healthMessage: { type: string, description: Readiness coverage or the reason it could not be determined. }
            message: { type: string }
            checkedAt: { type: string, format: date-time }
            lastSuccessfulCheckAt: { type: string, format: date-time }
            location: { type: string, const: Dispatch controller }
            resources:
              type: array
              items:
                type: object
                properties:
                  apiVersion: { type: string }
                  kind: { type: string }
                  namespace: { type: string }
                  name: { type: string }
                  state: { type: string, enum: [synced, out_of_sync, missing, unknown] }
                  health: { type: string }
                  message: { type: string, description: Sanitized resource access or readiness limitation. }
                  truncated: { type: boolean }
                  differences:
                    type: array
                    items:
                      type: object
                      properties:
                        path: { type: string }
                        expected: {}
                        actual: {}
                        redacted: { type: boolean }
        actions:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              appId: { type: string }
              deploymentId: { type: string }
              actor: { type: string }
              state: { type: string, enum: [running, succeeded, failed] }
              message: { type: string }
              createdAt: { type: string, format: date-time }
...
    Deployment:
      type: object
      required: [id, appId, commitSha, specDigest, state, message, createdAt]
      properties:
        id: { type: string }
        appId: { type: string }
        commitSha: { type: string }
        specDigest: { type: string }
        state:
          type: string
          enum: [queued, fetching, building, starting, checking, routing, succeeded, failed, cancelled]
        message: { type: string }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }